requirements.txt
boto3==1.35.98
botocore==1.35.98
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
idna==3.10
jmespath==1.0.1
loguru==0.7.3
numpy==2.2.1
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
pyarrow==18.1.0
pycodestyle==2.12.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
tzdata==2024.2
urllib3==2.3.0
win32_setctime==1.2.0
